% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMFPCA_1020_2WayNested.R
\name{LMFPCA_1020_2WayNested}
\alias{LMFPCA_1020_2WayNested}
\title{Perform Multilevel localized-variate functional principal component analysis (MLVFPCA)}
\usage{
LMFPCA_1020_2WayNested(data, m, L1, L2, option = list(SmoothD = "2Diff",
  rho_1 = NULL, rho1Seq = NULL, nsol = 10, FVE_M = 20, PrevPi = NULL,
  PrevPi_d = NULL, maxiter = 100, eps = 0.01, verbose = FALSE, k = 3, alpha
  = NULL, lambda = NULL, alphaSeq = NULL, lambdaSeq = NULL, FVE_threshold =
  0.85, nfold = 5, model = "2WayNested.spatial", levels = c(L1 = 100, L2 =
  5), rho1Seq.list = NULL, rho_1.list = NULL, alpha.list = NULL,
  lambda.list = NULL, alphaSeq.list = NULL, lambdaSeq.list = NULL, parallel
  = F, no_cores = 10, alphaproportion = NULL, c = NULL,      corr.rho =
  NULL, corr.far.quantile = 0.7))
}
\arguments{
\item{data}{a dataset where each row is individual time series that concatenates m length-p vectors. Each column contains L2 repeated measures nested within L1 subjects. The number of columns is pXm; and the number of rows is L1*L2.}

\item{m}{number of variates}

\item{L1}{number of subjects}

\item{L2}{the number of repeated measures within each subject}

\item{option}{K: the number of principal components to estimate in each level, default is 3; Rho_1.list: roughness parameter, if specified as NULL cross-validation will be used to select rho. default is NULL; Alpha.list: penalty parameter that controls the degree of among variate sparsity, if specified as NULL cross-validation or FVE will be used to select alpha, default is NULL; Lambda.list: penalty parameter that controls the degree of within variate sparsity, if specified as NULL cross-validation or FVE will be used to select alpha, default is NULL; alphaproportion: the fraction of variance one would like to sacrifice in return of interpretability using FVE method to select penalty parameters. If specified as a number, FVE method will be used; if specified as NULL, cross-validation will be used, default is NULL.}
}
\value{
See Details
}
\description{
About the methodology:
Multilevel localized-variate functional principal component analysis (MLVFPCA) extracts interpretable basis functions that account for intra- and inter-cluster variability in a clustered multivariate process. The methodology is motivated by popular neuroscience experiments where participant brain activity is recorded using modalities such as EEG or fMRI, summarized as power within multiple time-varying frequency bands in multiple brain regions. The basis functions found by MLVFPCA can be both localized within a variate (e.g. nonzero only within a subinterval of each frequency band) and sparse among variates (e.g. zero across an entire frequency band). The sparsity is achieved by a rank-one based convex optimization with matrix L1 and block Frobenius norm based penalties. To jointly model data across repeated measures (e.g. electrodes), the functional variability is decomposed into subject level and within subject level (e.g. electrode level) variability.
}
\details{
Output:
predx: L1XL2 by p*m matrix, is the predicted functional data for each observation within each subject; \cr
pred.margin: L1XL2 by p*m matrix, is the predicted marginal functional data for each observation; \cr
phi.w: p*m by k matrix, each column is the within subject level eigenvector for 1 to kth principal component; \cr
phi.z: p*m by k matrix, each column is the subject level eigenvector for 1 to kth principal component; \cr
PCscore.w: L1*L2 by k matrix, contains within subject level principal component score; \cr
PCscore.z:  L1*L2 by k matrix, contains subject level principal component score; \cr
FVE.w: k-element vector, is the fraction of variance explained by each PC of the within subject level; \cr
FVE.z: k-element vector, is the fraction of variance explained by each PC of the subject level; \cr
eigValue.w: k-element vector, contains within subject level eigenvalues; \cr
eigValue.w: k-element vector, contains subject level eigenvalues. \cr
}
